Seperated out cft authentication
