+++ src/main/routes/routes.ts
@@ -2,7 +2,6 @@ import { Application } from 'express';
 import { infoRequestHandler } from '@hmcts/info-provider';
 import os from 'os';
 import process from 'process';
-import config = require('config');
 import fileErrorHandlerMiddleware from '../middlewares/fileErrorHandler.middleware';
 import {
   isPermittedAdmin,
@@ -15,22 +14,12 @@ import {
   processAccountSignIn,
 } from '../authentication/authenticationHandler';
 import {SessionManagementService} from '../service/sessionManagementService';
-import {CFT_IDAM_URL, FRONTEND_URL} from "../helpers/envUrls";
-import axios from "axios";
-import jwt_decode from "jwt-decode";
-var querystring = require('querystring');
+import {processCftLogin} from "../authentication/cft-authentication";
 
 const passport = require('passport');
 const healthcheck = require('@hmcts/nodejs-healthcheck');
 const multer = require('multer');
 const sessionManagement = new SessionManagementService();
-
-let cftIdamClientSecret;
-if(process.env.CFT_IDAM_CLIENT_SECRET) {
-  cftIdamClientSecret = process.env.CFT_IDAM_CLIENT_SECRET;
-} else {
-  cftIdamClientSecret = config.get('secrets.pip-ss-kv.CFT_IDAM_CLIENT_SECRET') as string;
-}
 
 export default function(app: Application): void {
   const storage = multer.diskStorage({
@@ -94,36 +83,7 @@ export default function(app: Application): void {
   app.get('/admin-login', passport.authenticate('admin-login', { failureRedirect: '/'}), regenerateSession);
   app.get('/media-verification', passport.authenticate('media-verification', { failureRedirect: '/'}), regenerateSession);
   app.post('/login/return', forgotPasswordRedirect, passport.authenticate('login', { failureRedirect: '/view-option'}), processAccountSignIn);
-  app.get('/cft-login/return', async (req, res, next) => {
-
-    const params = {
-      client_id: 'app-pip-frontend',
-      client_secret: cftIdamClientSecret,
-      grant_type: 'authorization_code',
-      redirect_uri: FRONTEND_URL + '/cft-login/return',
-      code: req.query.code as string,
-    };
-
-    const tokenRequest = axios.create({baseURL: CFT_IDAM_URL, timeout: 10000});
-
-    try {
-      const response = await tokenRequest.post('/o/token', querystring.stringify(params), {
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/x-www-form-urlencoded'
-        }
-      });
-
-      const data = response.data;
-      const jwtToken = jwt_decode(data.id_token);
-
-      // @ts-ignore
-      req.session.user = jwtToken;
-    } catch (e) {
-      console.log(e);
-    }
-    next();
-  }, regenerateSession);
+  app.get('/cft-login/return', processCftLogin, processCftAccount, regenerateSession);
 
   app.post('/media-verification/return', forgotPasswordRedirect, passport.authenticate('media-verification', { failureRedirect: '/view-option'}),
     (_req, res) => {mediaVerificationHandling(_req, res);});
